preparations for x64 port

+++ win32/q_shwin.c
@@ -139,9 +139,9 @@ void Sys_Mkdir (char *path)
 
 //============================================
 
-char	findbase[MAX_OSPATH];
-char	findpath[MAX_OSPATH];
-int		findhandle;
+char		findbase[MAX_OSPATH];
+char		findpath[MAX_OSPATH];
+intptr_t	findhandle;
 
 static qboolean CompareAttributes( unsigned found, unsigned musthave, unsigned canthave )
 {
